requirements.txt
matplotlib
pyyaml
tensorflow==2.5